% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_grid_CV.R
\name{h2o.cv.rnd.grid.shrinkage}
\alias{h2o.cv.rnd.grid.shrinkage}
\title{h2o.cv.rnd.grid.shrinkage}
\usage{
h2o.cv.rnd.grid.shrinkage(dataset.h2o, listVars, split.perc = 0.6,
  model_algorithm = h2o.gbm, nfolds = 10, seq_shrinkage = c(0, seq(from =
  0.002, to = 0.008, by = 0.002), seq(from = 0.01, to = 0.2, by = 0.05)),
  model_params = list(shrinkage = 0.1, n.trees = 30, interaction.depth = 5,
  n.bins = 50, n.minobsinnode = 1000, distribution = "bernoulli", importance =
  FALSE, balance.classes = TRUE), gain.calculate = TRUE, gain.ngroups = 100,
  col.pred = 3, col.target = tail(h2o::colnames(dataset.h2o), n = 1))
}
\description{
It does a grid search for shrinkage.
\cr Check \code{\link{h2o.cv.rnd.opt.param}} for a parameter optimization that can work with different algorithms and differents parameters to optimize.
}
\seealso{
\code{\link{h2o.cv.rnd.grid.shrinkage}} \code{\link{h2o.cv.rnd.opt.param}} \code{\link{h2o.cv.rnd.opt.param.get_best}} \code{\link{h2o.ggplot.dfResultsCV}}
}

