% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_grid_CV.R
\name{h2o.cv.rnd.opt.param.get_best}
\alias{h2o.cv.rnd.opt.param.get_best}
\title{h2o.cv.rnd.opt.param.get_best}
\usage{
h2o.cv.rnd.opt.param.get_best(dfResultsCV, metric = "auc.roc.test.median")
}
\description{
Selects best optimized parameter. By default selects best median on test set.
\cr Input is the output data.frame from \code{\link{h2o.cv.rnd.opt.param}}.
}
\seealso{
\code{\link{h2o.cv.rnd.grid.shrinkage}} \code{\link{h2o.cv.rnd.opt.param}} \code{\link{h2o.cv.rnd.opt.param.get_best}} \code{\link{h2o.ggplot.dfResultsCV}}
}

