% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_grid_CV.R
\name{h2o.cv.rnd.opt.param}
\alias{h2o.cv.rnd.opt.param}
\title{h2o.cv.rnd.opt.param}
\usage{
h2o.cv.rnd.opt.param(dataset.h2o, split.perc = 0.6,
  model_algorithm = h2o.gbm, nfolds = 10, max_it = 10, opt.area = 0.5,
  opt.tol = 0.01, opt.var = "shrinkage", opt.val = 0.1, opt.val.min = 0,
  opt.val.max = 1, opt.val.is.integer = FALSE,
  opt.criteria = "overfitting", opt.val.rang.perc = 0.9,
  model_params = list(shrinkage = 0.1, n.trees = 30, interaction.depth = 5,
  n.bins = 50, n.minobsinnode = 1000, distribution = "bernoulli", importance =
  FALSE, balance.classes = TRUE), gain.calculate = TRUE, gain.ngroups = 100,
  col.pred = 3, col.target = tail(h2o::colnames(dataset.h2o), n = 1))
}
\description{
Optimizes 1 parameter. Algorithm tries to avoid over-fitting. So, it does not optimizes test results. Instead, it looks for overlapping of train and test metrics.
\cr Check \code{\link{h2o.cv.rnd.opt.param.get_best}} for select best parameter from output from this function.
\cr Check \code{\link{h2o.ggplot.dfResultsCV}} for ploting output results from this function.
}
\seealso{
\code{\link{h2o.cv.rnd.grid.shrinkage}} \code{\link{h2o.cv.rnd.opt.param}} \code{\link{h2o.cv.rnd.opt.param.get_best}} \code{\link{h2o.ggplot.dfResultsCV}}
}

