% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassificationHelper.R
\name{h2o.get_varimp}
\alias{h2o.get_varimp}
\title{h2o.get_varimp}
\usage{
h2o.get_varimp(algo, model, top_features = 5)
}
\description{
Offitial H2O helper function (binaryClassificationHelper.R).
}

