% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_grid_CV.R
\name{h2o.getAUC.Gain}
\alias{h2o.getAUC.Gain}
\title{h2o.getAUC.Gain}
\usage{
h2o.getAUC.Gain(model, dataset.h2o, ngroups = 100, col.pred = 3,
  col.target = tail(h2o::colnames(dataset.h2o), n = 1))
}

