% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_grid_CV.R
\name{h2o.getListModels}
\alias{h2o.getListModels}
\title{h2o.getListModels}
\usage{
h2o.getListModels(models.grid)
}
\description{
Returns an R list of models from different inputs. Input is normally a base h2o grid search.
}

