% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.categoric}
\alias{h2o.ggplot.categoric}
\title{h2o.ggplot.categoric}
\usage{
h2o.ggplot.categoric(var_cat, var_score, title = NULL, ylab = NULL,
  xlab = NULL, outlier.color = NA, shape = ".", alpha = 0.3,
  angle = 90, ylimit = NULL, legend.position = "none")
}

