% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_grid_CV.R
\name{h2o.ggplot.dfResultsCV}
\alias{h2o.ggplot.dfResultsCV}
\title{h2o.ggplot.dfResultsCV}
\usage{
h2o.ggplot.dfResultsCV(dfResultsCV, list.vars = c("auc.roc.test",
  "auc.roc.train"), title = NULL, ylab = "AUC", xlab = NULL,
  guides = FALSE, horizontal = FALSE, useFactor = TRUE)
}
\description{
Plots boxplots of some metrics. By default AUC of ROC Curve for train and test.
\cr Input is the output data.frame from \code{\link{h2o.cv.rnd.opt.param}}.
}
\seealso{
\code{\link{h2o.cv.rnd.grid.shrinkage}} \code{\link{h2o.cv.rnd.opt.param}} \code{\link{h2o.cv.rnd.opt.param.get_best}} \code{\link{h2o.ggplot.dfResultsCV}}
}

