% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.dfVarImp}
\alias{h2o.ggplot.dfVarImp}
\title{h2o.ggplot.dfVarImp}
\usage{
h2o.ggplot.dfVarImp(dfVarImp, max_vars = 20, range_vars = NULL,
  binwidth = 0.05, fill = rgb(0.85, 0.72, 0.2, 0.5), color = rgb(0, 0, 0,
  0.8), horizontal = TRUE, reverseOrder = FALSE, title = "Model",
  ylab = "Percent. Influence", xlab = NULL, angle = 65, hjust = 1,
  log = FALSE, description = "Distributed GBM")
}

