% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.distribution}
\alias{h2o.ggplot.distribution}
\title{h2o.ggplot.distribution}
\usage{
h2o.ggplot.distribution(h2o_parsed_data, log10scale = 2, angle = 65,
  order = FALSE, orderRev = FALSE, color = rgb(0.85, 0.72, 0.2, 0.5),
  stroke = rgb(0, 0, 0, 0.2), title = NULL, xlab = NULL, ylab = NULL,
  stat = "identity")
}

