% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.gain}
\alias{h2o.ggplot.gain}
\title{h2o.ggplot.gain}
\usage{
h2o.ggplot.gain(dfGains, title = "Cum. Gain", ylab = "\% responses",
  xlab = "\% customers", addAUC = TRUE, ndigits = 5, color = rgb(0.1,
  0.1, 0.6, 0.4), color.base = rgb(0, 0, 0, 0.2))
}

