% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.gains}
\alias{h2o.ggplot.gains}
\title{h2o.ggplot.gains}
\usage{
h2o.ggplot.gains(dfGains.train, dfGains.test, title = "Cum. Gain",
  ylab = "\% responses", xlab = "\% customers", addAUC = TRUE,
  ndigits = 5, color.train = rgb(0.6, 0.1, 0.1, 0.4),
  color.test = rgb(0.1, 0.1, 0.6, 0.4), color.point.train = rgb(0.7, 0.1,
  0.1, 0.7), color.point.test = rgb(0.1, 0.1, 0.7, 0.7), color.base = rgb(0,
  0, 0, 0.2))
}

