% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.numeric}
\alias{h2o.ggplot.numeric}
\title{h2o.ggplot.numeric}
\usage{
h2o.ggplot.numeric(var_num, var_score, smooth = "red",
  colors_grad = c("blue", "red"), alpha_smooth = 0.3, title = NULL,
  ylab = NULL, xlab = NULL, shape = ".", alpha = 0.1, xlimit = NULL,
  ylimit = NULL, legend.position = "none")
}

