% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.roc}
\alias{h2o.ggplot.roc}
\title{h2o.ggplot.roc}
\usage{
h2o.ggplot.roc(perf, title = "ROC Curve", ylab = "sensitivity (TPR)",
  xlab = "1 - specificity (FPR)", addAUC = TRUE, ndigits = 5,
  color = rgb(0.1, 0.1, 0.6, 0.4), color.point = rgb(0.1, 0.1, 0.7, 0.7),
  color.base = rgb(0, 0, 0, 0.2))
}

