% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.ggplot.rocs}
\alias{h2o.ggplot.rocs}
\title{h2o.ggplot.rocs}
\usage{
h2o.ggplot.rocs(perf.train, perf.test, title = "ROC Curve",
  ylab = "sensitivity (TPR)", xlab = "1 - specificity (FPR)",
  addAUC = TRUE, ndigits = 5, color.train = rgb(0.6, 0.1, 0.1, 0.4),
  color.test = rgb(0.1, 0.1, 0.6, 0.4), color.point.train = rgb(0.7, 0.1,
  0.1, 0.7), color.point.test = rgb(0.1, 0.1, 0.7, 0.7), color.base = rgb(0,
  0, 0, 0.2))
}

