% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassificationHelper.R
\name{h2o.leaderBoard}
\alias{h2o.leaderBoard}
\title{h2o.leaderBoard}
\usage{
h2o.leaderBoard(models, test_hex, response, top_features = 5)
}
\description{
Offitial H2O helper function (binaryClassificationHelper.R).
}

