% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_models.R
\name{h2o.performance.gaussian}
\alias{h2o.performance.gaussian}
\title{h2o.performance.gaussian}
\usage{
h2o.performance.gaussian(h2o.target, h2o.pred)
}
\description{
Calculates some metrics for regression models (gaussian).
}
\details{
h2o.performance.gaussian


This method returns some error metrics for regression models. It needs a couple of numeric arrays: the target and the prediction.
}

