% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_deprecated_plot.R
\name{h2o.plot.gain}
\alias{h2o.plot.gain}
\title{h2o.plot.gain}
\usage{
h2o.plot.gain(model_score, real_target, results, resultsAUC, model,
  strExportFile = NULL, width = 800, height = 800)
}

