% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_plot.R
\name{h2o.report.html}
\alias{h2o.report.html}
\title{h2o.report.html}
\usage{
h2o.report.html(openHTML = interactive(), folderHTML = tempdir(),
  tempFileHTML = tempfile(pattern = "h2o_report_html_", fileext = ".html",
  tmpdir = folderHTML), tempFileTXT = tempfile(pattern = "h2o_report_html_",
  fileext = ".txt", tmpdir = tempdir()))
}

