% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_grid_CV.R
\name{h2o.runif}
\alias{h2o.runif}
\title{h2o.runif}
\usage{
h2o.runif(dataset.h2o, min = 0, max = 1, seed = -1, useFile = TRUE)
}
\description{
It seems h2o.runif from h2o package always return the same vector. This function uses runif from R to gives same performance without that bug.
}

