% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_dataset.R
\name{h2o.setClassColumns}
\alias{h2o.setClassColumns}
\title{h2o.setClassColumns}
\usage{
h2o.setClassColumns(dataset.h2o, listVarsNumeric = param.dictionary.numeric,
  listVarsCategorical = param.dictionary.categorical,
  limit.num.messages = 30)
}

