% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_dataset.R
\name{h2o.setFactorColumns}
\alias{h2o.setFactorColumns}
\title{h2o.setFactorColumns}
\usage{
h2o.setFactorColumns(dataset.h2o,
  listVarsCategorical = param.dictionary.categorical,
  limit.num.messages = 30)
}

