% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_dataset.R
\name{h2o.setNumericColumns}
\alias{h2o.setNumericColumns}
\title{h2o.setNumericColumns}
\usage{
h2o.setNumericColumns(dataset.h2o, listVarsNumeric = param.dictionary.numeric,
  limit.num.messages = 30)
}

