% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryClassificationHelper.R
\name{h2o.validate}
\alias{h2o.validate}
\title{h2o.validate}
\usage{
h2o.validate(t0, model, modeltype, validation, response, varimp)
}
\description{
Offitial H2O helper function (binaryClassificationHelper.R).
}

