% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_base.R
\name{h2oInit}
\alias{h2oInit}
\title{h2oInit}
\usage{
h2oInit(h2oMEM = params$h2oMEM, h2oThreads = params$h2oThreads,
  h2oNAME = params$h2oNAME, h2oPORT = params$h2oPORT,
  h2oJAR = system.file("java", "h2o.jar", package = "h2o"),
  h2oOutput = "nohup.out", wait = FALSE, sleep = 10)
}

