% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_base.R
\name{h2oInitRemote}
\alias{h2oInitRemote}
\title{h2oInitRemote}
\usage{
h2oInitRemote(server_remote_h2o_ip = connData$server_remote_h2o_ip,
  server_remote_h2o_port = connData$server_remote_h2o_port,
  server_remote_h2o_user = connData$server_remote_h2o_user,
  server_remote_h2o_pass = connData$server_remote_h2o_pass,
  server_remote_h2o_java = connData$server_remote_h2o_java,
  server_remote_h2o_jar = connData$server_remote_h2o_jar,
  server_remote_h2o_folder = connData$server_remote_h2o_folder,
  server_remote_h2o_nohup = connData$server_remote_h2o_nohup,
  server_remote_h2o_nohup_file = connData$server_remote_h2o_nohup_file,
  h2oMEM = params$h2oMEM, h2oThreads = params$h2oThreads,
  h2oNAME = params$h2oNAME, h2oPORT = params$h2oPORT, sleep = 10)
}

