% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h2otools.R
\name{h2oModelAUC}
\alias{h2oModelAUC}
\title{h2oModelAUC}
\usage{
h2oModelAUC(ip, port, dataReal, keyReal, dataModel, keyModel,
  criterion = "maximum_F1")
}
\description{
AUC model evaluation
}
\details{
h2oModelAUC

This method opens a browser with html of AUC model evaluation.
}

