% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2oplot.R
\name{h2oPlotGain}
\alias{h2oPlotGain}
\title{h2oPlotGain}
\usage{
h2oPlotGain(model_score, real_target, results, resultsAUC, model,
  strExportFile = NULL, width = 800, height = 800)
}
\description{
h2oPlotGain
}

