% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h2oplot.R
\name{h2oPlotRelativeImportance}
\alias{h2oPlotRelativeImportance}
\title{h2oPlotRelativeImportance}
\usage{
h2oPlotRelativeImportance(model_score, real_target, results, resultsAUC, model,
  strExportFile = NULL, width = 800, height = 800)
}
\description{
h2oPlotRelativeImportance
}

