% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2oplot.R
\name{h2oPlotResults}
\alias{h2oPlotResults}
\title{h2oPlotResults}
\usage{
h2oPlotResults(model_score, real_target, results, resultsAUC, model,
  strExportFile = NULL, width = 800, height = 800)
}
\description{
h2oPlotResults
}

