% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h2omodels.R
\name{h2oSplitTrainValidateTest}
\alias{h2oSplitTrainValidateTest}
\title{h2oSplitTrainValidateTest}
\usage{
h2oSplitTrainValidateTest(datasetH2O, prob1, prob2)
}
\description{
h2oSplitTrainValidateTest
}

