% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2otools_base.R
\name{h2oToRDF}
\alias{h2oToRDF}
\title{H2OParsedData object to R data.frame object}
\usage{
h2oToRDF(h2oParsedData, use_hex_string = FALSE, stringsAsFactors = FALSE,
  blank.lines.skip = FALSE, ...)
}
\description{
\code{h2oToRDF} returns a R data.frame object.
}
\details{
h2oToRDF


Another way to achieve this seems to be: as.numeric(as.vector(as.data.frame(...)))
}
\seealso{
\code{\link{as.data.frame.H2OParsedData}}
}

