% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h2otools.R
\name{h2oUnlockKeys}
\alias{h2oUnlockKeys}
\title{h2oUnlockKeys}
\usage{
h2oUnlockKeys(h2oServer)
}
\description{
Unlock all keys
}
\details{
h2oUnlockKeys

This method is not safe. It is supposed to unlock all keys. Use at your own risk.
}

