% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2otools.R
\name{h2oWebStoreView}
\alias{h2oWebStoreView}
\title{h2oWebStoreView}
\usage{
h2oWebStoreView(ip = NULL, port = NULL, h2oServer = NULL)
}
\description{
Web Store View
}
\details{
h2oWebStoreView

Opens H2O url for Store View.
}

