% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h2otools.R
\name{h2oPerformanceAUC}
\alias{h2oPerformanceAUC}
\title{h2oPerformanceAUC}
\usage{
h2oPerformanceAUC(h2oServer, dataReal, keyReal, dataModel, keyModel,
  criterion = "maximum_F1")
}
\description{
AUC model performance evaluation
}
\details{
h2oPerformanceAUC

This method returns AUC model performance evaluation.
}

