% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h2otools.R
\name{h2oWebAUC}
\alias{h2oWebAUC}
\title{h2oWebAUC}
\usage{
h2oWebAUC(ip, port, dataReal, keyReal, dataModel, keyModel,
  criterion = "maximum_F1")
}
\description{
AUC model evaluation
}
\details{
h2oWebAUC

This method opens a browser with html of AUC model evaluation.
}

