% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h2omodels.R
\name{h2oWebAUC}
\alias{h2oWebAUC}
\title{h2oWebAUC}
\usage{
h2oWebAUC(dataReal, keyReal, dataModel, keyModel, criterion = "maximum_F1",
  ip = NULL, port = NULL, h2oServer = NULL)
}
\description{
Web AUC model evaluation
}
\details{
h2oWebAUC

Opens H2O url for AUC model evaluation.
}

